refactor: update tests namespace
